Change testing to focus on chronos bookkeeping
The previous testing approach with the timer_expiry module (now gone) actually tested the
Erlang timers a lot more than it tested the chronos functionality.

chronos has been refactored to have a chronos_command module that is mocked using EQC to
test that chronos works.

The testing of chronos' bookkeeping has been changed to use eqc_component, which makes it
a lot easier to understand what is going on.

In addition to the eqc_component test three tests have been added which covers that
chronos actually triggers things, that restarting a timer works as expected and that you
can cancel a timer and not see anything being triggered.

The test has been simplified to only start one server as it did not add anything but more
bookkeeping in the test.

A new `start/0` function has been added to allow starting nameless servers.

The README has been updated with sections on how to test with EQC mocking, linking
approach and a small road-map.

When a chronos server exists it will no longer cancel all the timers. This is not
necessary since all timers have the chroons server pid as the destination and will be
removed automatically when the process stops.